+++ libpsx-rs/src/types/register/b8_register.rs
@@ -54,12 +54,12 @@ impl B8Register {
 }
 
 impl B8MemoryMap for B8Register {
-    fn read_u8(&mut self, offset: usize) -> ReadResult<u8> {
+    fn read_u8(&mut self, offset: u32) -> ReadResult<u8> {
         if offset != 0 { panic!("Invalid offset"); }
         Ok(Self::read_u8(self))
     }
     
-    fn write_u8(&mut self, offset: usize, value: u8) -> WriteResult {
+    fn write_u8(&mut self, offset: u32, value: u8) -> WriteResult {
         if offset != 0 { panic!("Invalid offset"); }
         Self::write_u8(self, value);
         Ok(())
